+++ src/ff7Addresses.ts
@@ -92,6 +92,7 @@ export interface FF7Addresses {
   field_script_temp_vars: number;
   field_file_section1_ptr: number;
   field_models_objs: number;
+  field_line_objs: number;
 }
 
 export function useFF7Addresses() {